netdev-offload: e2e: Change signature of ct_counter_query offload API
Prestep for counter IDs commit. Change type of second parameter of
ct_counter_query offload API callback from uint32_t to uintptr_t.

Issue: 2333202

+++ lib/netdev-offload-provider.h
@@ -108,7 +108,7 @@ struct netdev_flow_api {
     const struct dpif_sflow_attr *(*sflow_attr_get)(uint32_t gid);
 
     /* Queries a CT counter object. */
-    int (*ct_counter_query)(struct netdev *, uint32_t, long long, long long,
+    int (*ct_counter_query)(struct netdev *, uintptr_t, long long, long long,
                             struct dpif_flow_stats *);
 
     /* Deinitializes the netdev flow api. */
